+++ cpukit/libcsupport/src/getpagesize.c
@@ -1,5 +1,5 @@
 /*
- *  $Id: getpagesize.c,v 1.1 2007/02/06 14:44:34 ralf Exp $
+ *  $Id: getpagesize.c,v 1.2 2009/09/24 11:14:12 joel Exp $
  */
 
 #if HAVE_CONFIG_H
@@ -9,16 +9,14 @@
 #include <unistd.h>
 #include <sys/param.h>
 
-/*PAGE
- *
+/*
  *  Get System Page Size (from SVR4 and 4.2+ BSD)
  *
  *  This is not a functional version but the SPARC backend for at least
  *  gcc 2.8.1 plus gnat 3.13p and gcc 3.0.1 require it to be there and
  *  return a reasonable value.
  */
-
-size_t getpagesize(void)
+int getpagesize(void)
 {
   return PAGE_SIZE;
 }
