+++ cpukit/score/src/pheapextend.c
@@ -1,4 +1,12 @@
 /**
+ * @file
+ *
+ * @ingroup ScoreProtHeap
+ *
+ * @brief Protected Heap Handler implementation.
+ */
+
+/*
  *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
@@ -6,7 +14,7 @@
  *  found in the file LICENSE in this distribution or at
  *  http://www.rtems.com/license/LICENSE.
  *
- *  $Id: pheapextend.c,v 1.2 2008/09/04 17:39:55 ralf Exp $
+ *  $Id: pheapextend.c,v 1.4 2009/09/06 15:24:07 joel Exp $
  */
 
 #if HAVE_CONFIG_H
@@ -19,11 +27,11 @@
 bool _Protected_heap_Extend(
   Heap_Control *the_heap,
   void         *starting_address,
-  size_t        size
+  uintptr_t     size
 )
 {
   Heap_Extend_status status;
-  uint32_t           amount_extended;
+  uintptr_t           amount_extended;
 
   _RTEMS_Lock_allocator();
     status = _Heap_Extend(the_heap, starting_address, size, &amount_extended);
